Refactor Presenter and IView with XML documentation
The `AddRegistrationPresenter.cs` file has been refactored to enhance readability and maintainability. The `Presenter` class is now public and includes XML documentation comments for its methods and constructors, providing better context for developers. The overall structure remains the same, but the code is more organized and self-explanatory. Additionally, the `IAddRegistrationView.cs` interface has been updated with XML documentation comments for each method and property, improving clarity on their purpose and usage.
